+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0 (2013-XX-XX)
+
+* Add the support of multiple managed files
+* Preserve other top-level keys than the configured one in the file
+* Add a rename map used to rename parameters when updating the parameters file
+* Add the possibility to use another top-level key than ``parameters``
+
 ## 2.0.0 (2013-04-06)
 
 * BC BREAK the env map has been changed, inverting the keys and the values. Refs #14